full token/admin securisation done

+++ controllers/audienceController.js
@@ -18,6 +18,10 @@ const getAudience = async(req, res) => {
 };
 
 const createAudience = async(req, res) => {
+    const userId = verifyToken(req, res);
+    if (userId == null) return res;
+    if(adminCheck(userId) == false) return res;
+    
     const target = req.body.target;
 
     const newAudienceRecord = new Audience({
@@ -34,6 +38,11 @@ const createAudience = async(req, res) => {
 };
 
 const updateAudience = async(req, res) => {
+
+    const userId = verifyToken(req, res);
+    if (userId == null) return res;
+    if(adminCheck(userId) == false) return res;
+
     const id = req.body.id;
     const target = req.body.target;
 
@@ -50,6 +59,11 @@ const updateAudience = async(req, res) => {
 };
 
 const deleteAudience = async(req, res) => {
+
+    const userId = verifyToken(req, res);
+    if (userId == null) return res;
+    if(adminCheck(userId) == false) return res;
+
     const id = req.body.id;
     if(!ObjectId.isValid(id)) return res.status(400).json({'error':'L\'ID spécifié n\'existe  pas'});
     Audience.findByIdAndDelete(id,(err, result) => {
@@ -59,6 +73,36 @@ const deleteAudience = async(req, res) => {
             return res.status(500).json({'error':'Erreur lors de la suppression'});
         }
     })
+}
+
+// checks is user retrieved from token has admin rights
+function adminCheck(userId) {
+    User.findById(userId, function (err, result) {
+        if(err) {
+            res.status(404).json({'error':'Utilisateur introuvable'});
+        } else {
+            if (result.isAdmin) {
+                return true;
+            } else {
+                res.status(403).json({'error':'Vous ne disposez pas des droits'});
+            }
+        }
+    })
+    return false;
+}
+
+// fonction de vérification de la validité du token, renvoie null si erreur
+function verifyToken(req, res) {
+    try {
+        jwt.verify(req.headers['authorization'], process.env.TOKEN_SECRET, function (tokenErr, decoded) {
+            if (tokenErr) throw new Error(tokenErr);
+            req.auth = decoded;
+        })
+    } catch (e) {
+       res.status(403).json({'error':'Token invalide '+e});
+       return null;
+    }
+    return req.auth.sub;
 }
 
 module.exports = { getAudience, createAudience, updateAudience, deleteAudience };
